+++ continuum/scenarios/transformation_incremental.py
@@ -5,7 +5,7 @@ from torchvision import transforms
 
 from continuum.datasets import _ContinuumDataset
 from continuum.scenarios import InstanceIncremental
-from continuum.task_set import TaskSet
+from continuum.tasks import TaskSet
 
 
 class TransformationIncremental(InstanceIncremental):
